requirements.txt
oauthlib==0.7.2
python-twitter==2.2
requests==2.5.1
requests-oauthlib==0.4.2
PyYAML==3.11
wsgiref==0.1.2
git+git://github.com/snare/scruffy.git